requirements.txt
streamlit>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
scipy>=1.11.0
groq>=0.4.0
